refactor: StationFixture를 domain패키지에서 fixture 패키지로 이
