--- src/test/java/nextstep/subway/member/domain/MemberFixture.java
@@ -1,5 +1,0 @@
-package nextstep.subway.member.domain;
-
-public class MemberFixture {
-
-}
